chore(gRPC): rename `utils.rs` to `macros.rs`
