+++ crates/iota-grpc-server/src/lib.rs
@@ -5,10 +5,10 @@
 pub mod constants;
 mod error;
 pub mod ledger_service;
+pub mod macros;
 pub mod server;
 pub mod transaction_execution_service;
 pub mod types;
-pub mod utils;
 
 // Re-export commonly used types and traits
 pub use ledger_service::LedgerGrpcService;
